passage au main.c => création des fonctioons loop et setup
