+++ ControleMoteur2.ioc
@@ -1,6 +1,6 @@
 #MicroXplorer Configuration settings - do not modify
-CAD.formats=
-CAD.pinconfig=
+CAD.formats=[]
+CAD.pinconfig=Dual
 CAD.provider=
 File.Version=6
 GPIO.groupedBy=Group By Peripherals
@@ -41,8 +41,8 @@ Mcu.UserConstants=
 Mcu.UserName=STM32L073RZTx
 MxCube.Version=6.13.0
 MxDb.Version=DB.6.0.130
-NVIC.EXTI0_1_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
-NVIC.EXTI2_3_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
+NVIC.EXTI0_1_IRQn=true\:0\:0\:false\:true\:true\:1\:true\:true\:true
+NVIC.EXTI2_3_IRQn=true\:0\:0\:false\:true\:true\:2\:true\:true\:true
 NVIC.ForceEnableDMAVector=true
 NVIC.HardFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
 NVIC.NonMaskableInt_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
@@ -73,8 +73,12 @@ PA5.GPIOParameters=GPIO_Label
 PA5.GPIO_Label=LD2 [Green Led]
 PA5.Locked=true
 PA5.Signal=GPIO_Output
+PC0.GPIOParameters=GPIO_PuPd
+PC0.GPIO_PuPd=GPIO_PULLUP
 PC0.Locked=true
 PC0.Signal=GPXTI0
+PC1.GPIOParameters=GPIO_PuPd
+PC1.GPIO_PuPd=GPIO_PULLUP
 PC1.Locked=true
 PC1.Signal=GPXTI1
 PC10.Locked=true
@@ -90,8 +94,12 @@ PC14-OSC32_IN.Signal=RCC_OSC32_IN
 PC15-OSC32_OUT.Locked=true
 PC15-OSC32_OUT.Mode=LSE-External-Oscillator
 PC15-OSC32_OUT.Signal=RCC_OSC32_OUT
+PC2.GPIOParameters=GPIO_PuPd
+PC2.GPIO_PuPd=GPIO_PULLUP
 PC2.Locked=true
 PC2.Signal=GPXTI2
+PC3.GPIOParameters=GPIO_PuPd
+PC3.GPIO_PuPd=GPIO_PULLUP
 PC3.Locked=true
 PC3.Signal=GPXTI3
 PC6.Locked=true
